requirements.txt
# python3 -m pip install -r requirements.txt
python-dotenv
pyyaml
matplotlib
langchain
langchain_openai
chainlit